+++ src/Sentry/SentrySdk.cs
@@ -18,7 +18,7 @@ namespace Sentry
     /// It allows safe static access to a client and scope management.
     /// When the SDK is uninitialized, calls to this class result in no-op so no callbacks are invoked.
     /// </remarks>
-    public static class SentrySdk
+    public static partial class SentrySdk
     {
         private static IHub _hub = DisabledHub.Instance;
 
@@ -41,9 +41,7 @@ namespace Sentry
             // from anywhere else, return a disabled hub.
             if (Dsn.IsDisabled(dsn))
             {
-                options.LogWarning(
-                    "Init was called but no DSN was provided nor located. Sentry SDK will be disabled.");
-
+                options.LogWarning("Init was called but no DSN was provided nor located. Sentry SDK will be disabled.");
                 return DisabledHub.Instance;
             }