+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 
 - Collect and send Client Reports to Sentry, which contain counts of discarded events. ([#1556](https://github.com/getsentry/sentry-dotnet/pull/1556))
 - Expose `ITransport` and `SentryOptions.Transport` public, to support using custom transports ([#1602](https://github.com/getsentry/sentry-dotnet/pull/1602))
+- Android native crash support ([#1288](https://github.com/getsentry/sentry-dotnet/pull/1288))
 
 ### Fixes